+++ klarinet-coroutines/src/commonTest/kotlin/com/vectencia/klarinet/coroutines/FileReaderExtensionsTest.kt
@@ -1,0 +1,11 @@
+package com.vectencia.klarinet.coroutines
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class FileReaderExtensionsTest {
+    @Test
+    fun fileReaderExtensionsCompile() {
+        assertTrue(true)
+    }
+}